Splits user view out and adds template
