+++ bookwyrm/templates/preferences/blocks.html
@@ -1,4 +1,4 @@
-{% extends 'preferences/preferences_layout.html' %}
+{% extends 'preferences/layout.html' %}
 {% load i18n %}
 
 {% block title %}{% trans "Blocked Users" %}{{ author.name }}{% endblock %}
