+++ bookwyrm/templates/preferences/change_password.html
@@ -1,4 +1,4 @@
-{% extends 'preferences/preferences_layout.html' %}
+{% extends 'preferences/layout.html' %}
 {% load i18n %}
 
 {% block title %}{% trans "Change Password" %}{% endblock %}
