+++ bookwyrm/templates/preferences/delete_account.html
@@ -1,0 +1,27 @@
+{% extends 'preferences/layout.html' %}
+{% load i18n %}
+
+{% block title %}{% trans "Delete Account" %}{% endblock %}
+
+{% block header %}
+{% trans "Delete Account" %}
+{% endblock %}
+
+{% block panel %}
+{% if form.non_field_errors %}
+<p class="notification is-danger">{{ form.non_field_errors }}</p>
+{% endif %}
+<div class="block">
+    <h2 class="title is-4">{% trans "Permanently delete account" %}</h2>
+    <p class="help">{% trans "Deleting your account cannot be undone. The username will not be available to register in the future." %}</p>
+
+    <form name="delete-account" action="/delete-account" method="post">
+        <div class="field">
+            <label class="label" for="id_password">{% trans "Confirm password" %}</label>
+            <input class="input" type="password" name="password" id="id_password">
+        </div>
+        <button type="submit" class="is-danger">Delete</button>
+    </form>
+</div>
+{% endblock %}
+
