+++ bookwyrm/templates/preferences/edit_user.html
@@ -1,4 +1,4 @@
-{% extends 'preferences/preferences_layout.html' %}
+{% extends 'preferences/layout.html' %}
 {% load i18n %}
 
 {% block title %}{% trans "Edit Profile" %}{% endblock %}
@@ -69,4 +69,3 @@
     </div>
     <div class="block"><button class="button is-primary" type="submit">{% trans "Save" %}</button></div>
 </form>
-{% endblock %}
