+++ bookwyrm/templates/preferences/layout.html
@@ -18,6 +18,10 @@
                 {% url 'prefs-password' as url %}
                 <a href="{{ url }}"{% if url in request.path %} class="is-active" aria-selected="true"{% endif %}>{% trans "Change Password" %}</a>
             </li>
+            <li>
+                {% url 'prefs-delete' as url %}
+                <a href="{{ url }}"{% if url in request.path %} class="is-active" aria-selected="true"{% endif %}>{% trans "Delete Account" %}</a>
+            </li>
         </ul>
         <h2 class="menu-label">{% trans "Relationships" %}</h2>
         <ul class="menu-list">
